+++ HandlingEditor.Client/HandlingEditor.Client.csproj
@@ -51,6 +51,7 @@
     <Compile Include="Config.cs" />
     <Compile Include="HandlingInfo.cs" />
     <Compile Include="HandlingEditor.cs" />
+    <Compile Include="HandlingMenu.cs" />
     <Compile Include="HandlingPreset.cs" />
     <Compile Include="Helpers.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
