+++ HandlingEditor.Client/HandlingEditor.cs
@@ -2,9 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Drawing;
 using System.Text;
-using NativeUI;
 using CitizenFX.Core;
 using CitizenFX.Core.UI;
 using static CitizenFX.Core.Native.API;
@@ -14,623 +12,74 @@ namespace HandlingEditor.Client
 {
     public class HandlingEditor : BaseScript
     {
-        private static string ResourceName;
-        private static readonly string ScriptName = "Handling Editor";
-        private static readonly string kvpPrefix = "handling_";
+        #region EVENTS
 
-        #region CONFIG_FIEDS
-        private static float FloatPrecision = 0.001f;
-        private static float FloatStep = 0.01f;
-        private static float ScriptRange = 150.0f;
-        private static long timer = 1000;
-        private static bool debug = false;
-        private static int toggleMenu = 168;
-        private static float screenPosX = 1.0f;
-        private static float screenPosY = 0.0f;
+        public static event EventHandler InvalidatedGUI;
+
+        #endregion
+
+        #region CONFIG FIEDS
+        public static float FloatPrecision { get; private set; } = 0.001f;
+        public static float FloatStep { get; private set; } = 0.01f;
+        public static float ScriptRange { get; private set; } = 150.0f;
+        public static long Timer { get; private set; } = 1000;
+        public static bool Debug { get; private set; } = false;
+        public static int ToggleMenu { get; private set; } = 168;
+        public static float ScreenPosX { get; private set; } = 1.0f;
+        public static float ScreenPosY { get; private set; } = 0.0f;
         #endregion
 
         #region FIELDS
-        private HandlingInfo handlingInfo;
-        private Dictionary<string,HandlingPreset> serverPresets;
-        private long currentTime;
-        private long lastTime;
-        private int playerPed;
-        private int currentVehicle;
-        private HandlingPreset currentPreset;
-        private IEnumerable<int> vehicles;
+        public const string ScriptName = "Handling Editor";
+        public const string kvpPrefix = "handling_";
+        public static string ResourceName { get; private set; }
+        public static HandlingInfo HandlingInfo { get; private set; }
+        public static Dictionary<string,HandlingPreset> ServerPresets { get; private set; }
+        public static long CurrentTime { get; private set; }
+        public static long LastTime { get; private set; }
+        public static int PlayerPed { get; private set; }
+        public static int CurrentVehicle { get; private set; }
+        public static HandlingPreset CurrentPreset { get; private set; }
+        public static IEnumerable<int> Vehicles { get; private set; }
         #endregion
 
-        #region GUI_FIELDS
-        private MenuPool _menuPool;
-        private UIMenu EditorMenu;
-        private UIMenu PersonalPresetsMenu;
-        private UIMenu ServerPresetsMenu;
-        #endregion
-
-        #region GUI_METHODS
-        private async Task<string> GetOnScreenString(string defaultText)
-        {
-            DisableAllControlActions(1);
-            AddTextEntry("ENTER_VALUE", "Enter value");
-            DisplayOnscreenKeyboard(1, "ENTER_VALUE", "", defaultText, "", "", "", 128);
-            while (UpdateOnscreenKeyboard() != 1 && UpdateOnscreenKeyboard() != 2) await Delay(100);
-            EnableAllControlActions(1);
-            return GetOnscreenKeyboardResult();
-        }
-
-        private UIMenuDynamicListItem AddDynamicFloatList(UIMenu menu, FieldInfo<float> fieldInfo)
-        {
-            string name = fieldInfo.Name;
-            string description = fieldInfo.Description;
-            float min = fieldInfo.Min;
-            float max = fieldInfo.Max;
-
-            if (!currentPreset.Fields.ContainsKey(name))
-                return null;
-
-            float value = currentPreset.Fields[name];
-            string FloatChangeCallBack(UIMenuDynamicListItem item, UIMenuDynamicListItem.ChangeDirection direction)
-            {
-                if (direction == UIMenuDynamicListItem.ChangeDirection.Left)
-                {
-                    var newvalue = value - FloatStep;
-                    if (newvalue < min)
-                        Screen.ShowNotification($"{ScriptName}: Min value allowed for ~b~{name}~w~ is {min}");
-                    else
-                    {
-                        value = newvalue;
-                        currentPreset.Fields[name] = newvalue;
-                    }
-                }
-                else
-                {
-                    var newvalue = value + FloatStep;
-                    if (newvalue > max)
-                        Screen.ShowNotification($"{ScriptName}: Max value allowed for ~b~{name}~w~ is {max}");
-                    else
-                    {
-                        value = newvalue;
-                        currentPreset.Fields[name] = newvalue;
-                    }
-                }
-                return value.ToString("F3");
-            };
-            var newitem = new UIMenuDynamicListItem(name, description, value.ToString("F3"), FloatChangeCallBack);
-
-            menu.AddItem(newitem);
-
-            EditorMenu.OnItemSelect += async (sender, item, index) =>
-            {
-                if (item == newitem)
-                {
-                    EditorMenu.Visible = false;
-
-                    string text = await GetOnScreenString(value.ToString());
-                    float newvalue = value;
-
-                    if (float.TryParse(text, out newvalue))
-                    {
-                        if(newvalue >= min && newvalue <= max)
-                            currentPreset.Fields[name] = newvalue;
-                        else
-                            Screen.ShowNotification($"{ScriptName}:  Value out of allowed limits for ~b~{name}~w~, Min:{min}, Max:{max}");
-                    }else
-                        Screen.ShowNotification($"{ScriptName}:  Invalid value for ~b~{name}~w~");
-
-                    int currentSelection = EditorMenu.CurrentSelection;
-                    UpdateEditorMenu(); //Should just update the current item instead
-                    EditorMenu.CurrentSelection = currentSelection;
-                    EditorMenu.Visible = true;
-                }
-            };
-
-            return newitem;
-        }
-
-        private UIMenuDynamicListItem AddDynamicIntList(UIMenu menu, FieldInfo<int> fieldInfo)
-        {
-            string name = fieldInfo.Name;
-            string description = fieldInfo.Description;
-            int min = fieldInfo.Min;
-            int max = fieldInfo.Max;
-
-            if (!currentPreset.Fields.ContainsKey(name))
-                return null;
-
-            int value = currentPreset.Fields[name];
-            string IntChangeCallBack(UIMenuDynamicListItem item, UIMenuDynamicListItem.ChangeDirection direction)
-            {
-                if (direction == UIMenuDynamicListItem.ChangeDirection.Left)
-                {
-                    var newvalue = value - 1;
-                    if (newvalue < min)
-                        Screen.ShowNotification($"{ScriptName}: Min value allowed for ~b~{name}~w~ is {min}");
-                    else
-                    {
-                        value = newvalue;
-                        currentPreset.Fields[name] = newvalue;
-                    }
-                }
-                else
-                {
-                    var newvalue = value + 1;
-                    if (newvalue > max)
-                        Screen.ShowNotification($"{ScriptName}: Max value allowed for ~b~{name}~w~ is {max}");
-                    else
-                    {
-                        value = newvalue;
-                        currentPreset.Fields[name] = newvalue;
-                    }
-                }
-                return value.ToString();
-            };
-            var newitem = new UIMenuDynamicListItem(name, description, value.ToString(), IntChangeCallBack);
-
-            menu.AddItem(newitem);
-
-            EditorMenu.OnItemSelect += async (sender, item, index) =>
-            {
-                if (item == newitem)
-                {
-                    EditorMenu.Visible = false;
-
-                    string text = await GetOnScreenString(value.ToString());
-                    int newvalue = value;
-
-                    if (int.TryParse(text, out newvalue))
-                    {
-                        if (newvalue >= min && newvalue <= max)
-                            currentPreset.Fields[name] = newvalue;
-                        else
-                            Screen.ShowNotification($"{ScriptName}: Value out of allowed limits for ~b~{name}~w~, Min:{min}, Max:{max}");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: Invalid value for ~b~{name}~w~");
-
-                    int currentSelection = EditorMenu.CurrentSelection;
-                    UpdateEditorMenu();  //Should just update the current item instead
-                    EditorMenu.CurrentSelection = currentSelection;
-                    EditorMenu.Visible = true;
-                }
-            };
-
-            return newitem;
-        }
-
-        private UIMenuDynamicListItem[] AddDynamicVector3List(UIMenu menu, FieldInfo<Vector3> fieldInfo)
-        {
-            string fieldName = fieldInfo.Name;
-
-            if (!currentPreset.Fields.ContainsKey(fieldName))
-                return null;
-
-            string fieldDescription = fieldInfo.Description;
-            Vector3 fieldMin = fieldInfo.Min;
-            Vector3 fieldMax = fieldInfo.Max;
-
-            string fieldNameX = $"{fieldName}_x";
-            float valueX = currentPreset.Fields[fieldName].X;
-            float minValueX = fieldMin.X;
-            float maxValueX = fieldMax.X;
-
-            var newitemX = new UIMenuDynamicListItem(fieldNameX, fieldDescription, valueX.ToString("F3"), (sender, direction) =>
-            {
-                if (direction == UIMenuDynamicListItem.ChangeDirection.Left)
-                {
-                    var newvalue = valueX - FloatStep;
-                    if (newvalue < minValueX)
-                        Screen.ShowNotification($"{ScriptName}: Min value allowed for ~b~{fieldNameX}~w~ is {minValueX}");
-                    else
-                    {
-                        valueX = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].X = newvalue;
-                    }
-                }
-                else
-                {
-                    var newvalue = valueX + FloatStep;
-                    if (newvalue > maxValueX)
-                        Screen.ShowNotification($"{ScriptName}: Max value allowed for ~b~{fieldNameX}~w~ is {maxValueX}");
-                    else
-                    {
-                        valueX = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].X = newvalue;
-                    }
-                }
-                return valueX.ToString("F3");
-            });
-
-            menu.AddItem(newitemX);
-
-            string fieldNameY = $"{fieldName}_y";
-            float valueY = currentPreset.Fields[fieldName].Y;
-            float minValueY = fieldMin.Y;
-            float maxValueY = fieldMax.Y;
-
-            var newitemY = new UIMenuDynamicListItem(fieldNameY, fieldDescription, valueY.ToString("F3"), (sender, direction) =>
-            {
-                if (direction == UIMenuDynamicListItem.ChangeDirection.Left)
-                {
-                    var newvalue = valueY - FloatStep;
-                    if (newvalue < minValueY)
-                        Screen.ShowNotification($"{ScriptName}: Min value allowed for ~b~{fieldNameY}~w~ is {minValueY}");
-                    else
-                    {
-                        valueY = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].Y = newvalue;
-                    }
-                }
-                else
-                {
-                    var newvalue = valueY + FloatStep;
-                    if (newvalue > maxValueY)
-                        Screen.ShowNotification($"{ScriptName}: Max value allowed for ~b~{fieldNameY}~w~ is {maxValueY}");
-                    else
-                    {
-                        valueY = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].Y = newvalue;
-                    }
-                }
-                return valueY.ToString("F3");
-            });
-
-            menu.AddItem(newitemY);
-
-            string fieldNameZ = $"{fieldName}_z";
-            float valueZ = currentPreset.Fields[fieldName].Z;
-            float minValueZ = fieldMin.Z;
-            float maxValueZ = fieldMax.Z;
-
-            var newitemZ = new UIMenuDynamicListItem(fieldNameZ, fieldDescription, valueZ.ToString("F3"), (sender, direction) =>
-            {
-                if (direction == UIMenuDynamicListItem.ChangeDirection.Left)
-                {
-                    var newvalue = valueZ - FloatStep;
-                    if (newvalue < minValueZ)
-                        Screen.ShowNotification($"{ScriptName}: Min value allowed for ~b~{fieldNameZ}~w~ is {minValueZ}");
-                    else
-                    {
-                        valueZ = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].Z = newvalue;
-                    }
-                }
-                else
-                {
-                    var newvalue = valueZ + FloatStep;
-                    if (newvalue > maxValueZ)
-                        Screen.ShowNotification($"{ScriptName}: Max value allowed for ~b~{fieldNameZ}~w~ is {maxValueZ}");
-                    else
-                    {
-                        valueZ = newvalue;
-                        currentPreset.Fields[fieldInfo.Name].Z = newvalue;
-                    }
-                }
-                return valueZ.ToString("F3");
-            });
-
-            menu.AddItem(newitemZ);
-
-            EditorMenu.OnItemSelect += async (sender, item, index) =>
-            {
-                if (item == newitemX)
-                {
-                    EditorMenu.Visible = false;
-
-                    string text = await GetOnScreenString(valueX.ToString());
-                    float newvalue = valueX;
-
-                    if (float.TryParse(text, out newvalue))
-                    {
-                        if (newvalue >= minValueX && newvalue <= maxValueX)
-                            currentPreset.Fields[fieldName].X = newvalue;
-                        else
-                            Screen.ShowNotification($"{ScriptName}: Value out of allowed limits for ~b~{fieldNameX}~w~, Min:{minValueX}, Max:{maxValueX}");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: Invalid value for ~b~{fieldNameX}~w~");
-
-                    int currentSelection = EditorMenu.CurrentSelection;
-                    UpdateEditorMenu();  //Should just update the current item instead
-                    EditorMenu.CurrentSelection = currentSelection;
-                    EditorMenu.Visible = true;
-                }
-                else if (item == newitemY)
-                {
-                    EditorMenu.Visible = false;
-
-                    string text = await GetOnScreenString(valueY.ToString());
-                    float newvalue = valueY;
-
-                    if (float.TryParse(text, out newvalue))
-                    {
-                        if (newvalue >= minValueY && newvalue <= maxValueY)
-                            currentPreset.Fields[fieldName].Y = newvalue;
-                        else
-                            Screen.ShowNotification($"{ScriptName}: Value out of allowed limits for ~b~{fieldNameY}~w~, Min:{minValueY}, Max:{maxValueY}");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: Invalid value for ~b~{fieldNameY}~w~");
-
-                    int currentSelection = EditorMenu.CurrentSelection;
-                    UpdateEditorMenu();  //Should just update the current item instead
-                    EditorMenu.CurrentSelection = currentSelection;
-                    EditorMenu.Visible = true;
-                }
-                else if (item == newitemZ)
-                {
-                    EditorMenu.Visible = false;
-
-                    string text = await GetOnScreenString(valueZ.ToString());
-                    float newvalue = valueZ;
-
-                    if (float.TryParse(text, out newvalue))
-                    {
-                        if (newvalue >= minValueZ && newvalue <= maxValueZ)
-                            currentPreset.Fields[fieldName].Z = newvalue;
-                        else
-                            Screen.ShowNotification($"{ScriptName}: Value out of allowed limits for ~b~{fieldNameZ}~w~, Min:{minValueZ}, Max:{maxValueZ}");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: Invalid value for ~b~{fieldNameZ}~w~");
-
-                    int currentSelection = EditorMenu.CurrentSelection;
-                    UpdateEditorMenu();  //Should just update the current item instead
-                    EditorMenu.CurrentSelection = currentSelection;
-                    EditorMenu.Visible = true;
-                }
-            };
-
-            return new UIMenuDynamicListItem[3] { newitemX, newitemY, newitemZ };
-        }
-
-        private UIMenuItem AddLockedItem(UIMenu menu, BaseFieldInfo fieldInfo)
-        {
-            var newitem = new UIMenuItem(fieldInfo.Name, fieldInfo.Description);
-            newitem.Enabled = false;
-            newitem.SetRightBadge(UIMenuItem.BadgeStyle.Lock);
-
-            menu.AddItem(newitem);
-
-            menu.OnItemSelect += (sender, item, index) =>
-            {
-                if (item == newitem)
-                {
-                    Screen.ShowNotification($"{ScriptName}: The server doesn't allow to edit this field.");
-                }
-            };
-            return newitem;
-        }
-
-        private void UpdatePersonalPresetsMenu()
-        {
-            if (PersonalPresetsMenu == null)
-                return;
-
-            PersonalPresetsMenu.Clear();
-            KvpEnumerable kvpList = new KvpEnumerable(kvpPrefix);
-            foreach (var key in kvpList)
-            {
-                string value = GetResourceKvpString(key);
-                PersonalPresetsMenu.AddItem(new UIMenuItem(key.Remove(0, kvpPrefix.Length)));
-            }
-
-            PersonalPresetsMenu.RefreshIndex();
-        }
-
-        private void UpdateServerPresetsMenu()
-        {
-            if (ServerPresetsMenu == null)
-                return;
-
-            ServerPresetsMenu.Clear();
-
-            foreach (var preset in serverPresets)
-                ServerPresetsMenu.AddItem(new UIMenuItem(preset.Key));
-
-            ServerPresetsMenu.RefreshIndex();
-        }
-
-        private void UpdateEditorMenu()
-        {
-            if (EditorMenu == null)
-                return;
-
-            EditorMenu.Clear();
-
-            var PersonalPresetsItem = new UIMenuItem("Personal Presets", "The handling presets saved by you.");
-            PersonalPresetsItem.SetRightLabel("→→→");
-            EditorMenu.AddItem(PersonalPresetsItem);
-            EditorMenu.BindMenuToItem(PersonalPresetsMenu, PersonalPresetsItem);
-
-            var ServerPresetsItem = new UIMenuItem("Server Presets", "The handling presets loaded from the server.");
-            ServerPresetsItem.SetRightLabel("→→→");
-            EditorMenu.AddItem(ServerPresetsItem);
-            EditorMenu.BindMenuToItem(ServerPresetsMenu, ServerPresetsItem);
-
-            foreach (var item in handlingInfo.FieldsInfo)
-            {
-                var fieldInfo = item.Value;
-
-                if (fieldInfo.Editable)
-                {
-                    
-                    //string fieldName = fieldInfo.Name;
-                    //string fieldDescription = fieldInfo.Description;
-                    Type fieldType = fieldInfo.Type;
-
-                    if (fieldType == FieldType.FloatType)
-                        AddDynamicFloatList(EditorMenu, (FieldInfo<float>)item.Value);
-                    else if (fieldType == FieldType.IntType)
-                        AddDynamicIntList(EditorMenu, (FieldInfo<int>)item.Value);
-                    else if (fieldType == FieldType.Vector3Type)
-                        AddDynamicVector3List(EditorMenu, (FieldInfo<Vector3>)item.Value);
-                }
-                else
-                {
-                    //AddLockedItem(EditorMenu, item.Value);
-                }
-            }
-
-            var resetItem = new UIMenuItem("Reset", "Restores the default values");
-            EditorMenu.AddItem(resetItem);
-
-            EditorMenu.OnItemSelect += async (sender, item, index) =>
-            {
-                if (item == resetItem)
-                {
-                    currentPreset.Reset();
-                    RefreshVehicleUsingPreset(currentVehicle, currentPreset);
-                    RemoveDecorators(currentVehicle);
-
-                    await Delay(100);
-                    UpdateEditorMenu();
-                    EditorMenu.Visible = true;
-                }
-            };
-
-            UpdatePersonalPresetsMenu();
-            UpdateServerPresetsMenu();
-
-            EditorMenu.RefreshIndex();
-        }
-
-        private void InitialiseMenu()
-        {
-            if (EditorMenu == null)
-            {
-                EditorMenu = new UIMenu(ScriptName, "Editor", new PointF(screenPosX * Screen.Width, screenPosY * Screen.Height))
-                {
-                    MouseEdgeEnabled = false,
-                    ControlDisablingEnabled = false,
-                    MouseControlsEnabled = false
-                };
-            }
-            if (PersonalPresetsMenu == null)
-            {
-                PersonalPresetsMenu = new UIMenu(ScriptName, "Personal Presets")
-                {
-                    MouseEdgeEnabled = false,
-                    ControlDisablingEnabled = false,
-                    MouseControlsEnabled = false,
-                };
-
-                PersonalPresetsMenu.AddInstructionalButton(new InstructionalButton(Control.PhoneExtraOption, GetLabelText("ITEM_SAVE")));
-                PersonalPresetsMenu.AddInstructionalButton(new InstructionalButton(Control.PhoneOption, GetLabelText("ITEM_DEL")));
-            }
-            if (ServerPresetsMenu == null)
-            {
-                ServerPresetsMenu = new UIMenu(ScriptName, "Server Presets")
-                {
-                    MouseEdgeEnabled = false,
-                    ControlDisablingEnabled = false,
-                    MouseControlsEnabled = false,
-                };
-            }
-
-            UpdateEditorMenu();
-
-            PersonalPresetsMenu.OnItemSelect += (sender, item, index) =>
-            {
-                if(sender == PersonalPresetsMenu)
-                {
-                    string key = $"{kvpPrefix}{item.Text}";
-                    string value = GetResourceKvpString(key);
-                    if (value != null)
-                    {
-                        value = Helpers.RemoveByteOrderMarks(value);
-                        XmlDocument doc = new XmlDocument();
-                        doc.LoadXml(value);
-                        var handling = doc["Item"];
-                        GetPresetFromXml(handling, currentPreset);
-
-                        int currentSelection = PersonalPresetsMenu.CurrentSelection;
-                        UpdateEditorMenu();
-                        PersonalPresetsMenu.CurrentSelection = currentSelection;
-                        Screen.ShowNotification($"{ScriptName}: Personal preset ~b~{item.Text}~w~ applied");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: ~r~ERROR~w~ Personal preset ~b~{item.Text}~w~ corrupted");
-                }
-            };
-
-            ServerPresetsMenu.OnItemSelect += (sender, item, index) =>
-            {
-                if(sender == ServerPresetsMenu)
-                {
-                    string key = item.Text;
-                    if (serverPresets.ContainsKey(key))
-                    {
-                        foreach (var field in serverPresets[key].Fields.Keys)
-                        {
-                            if (currentPreset.Fields.ContainsKey(field))
-                            {
-                                currentPreset.Fields[field] = serverPresets[key].Fields[field];
-                            }
-                            else Debug.Write($"Missing {field} field in currentPreset");
-                        }
-
-                        int currentSelection = PersonalPresetsMenu.CurrentSelection;
-                        UpdateEditorMenu();
-                        PersonalPresetsMenu.CurrentSelection = currentSelection;
-                        Screen.ShowNotification($"{ScriptName}: Server preset ~b~{key}~w~ applied");
-                    }
-                    else
-                        Screen.ShowNotification($"{ScriptName}: ~r~ERROR~w~ Server preset ~b~{key}~w~ corrupted");
-                }
-            };
-
-            if (_menuPool == null)
-            {
-                _menuPool = new MenuPool()
-                {
-                    ResetCursorOnOpen = true
-                };
-            }
-            else _menuPool.ToList().Clear();
-            
-            _menuPool.Add(EditorMenu);
-            _menuPool.Add(PersonalPresetsMenu);
-            _menuPool.Add(ServerPresetsMenu);
-
-            _menuPool.RefreshIndex();
-        }
-        #endregion
+        #region CONSTRUCTOR
 
         public HandlingEditor()
         {
             ResourceName = GetCurrentResourceName();
-            Debug.WriteLine($"{ScriptName}: Script by Neos7");
+            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Script by Neos7");
             LoadConfig();
             
-            handlingInfo = new HandlingInfo();
+            HandlingInfo = new HandlingInfo();
             ReadFieldInfo();
-            serverPresets = new Dictionary<string, HandlingPreset>();
+            ServerPresets = new Dictionary<string, HandlingPreset>();
             ReadServerPresets();
 
             RegisterDecorators();
 
             ReadVehiclePermissions();
 
-            currentTime = GetGameTimer();
-            lastTime = currentTime;
-            currentPreset = null;
-            currentVehicle = -1;
-            vehicles = Enumerable.Empty<int>();
+            CurrentTime = GetGameTimer();
+            LastTime = CurrentTime;
+            CurrentPreset = null;
+            CurrentVehicle = -1;
+            Vehicles = Enumerable.Empty<int>();
 
             RegisterCommand("handling_range", new Action<int, dynamic>((source, args) =>
             {
                 if(args.Count < 1)
                 {
-                    Debug.WriteLine($"{ScriptName}: Missing float argument");
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Missing float argument");
                     return;
                 }
 
                 if (float.TryParse(args[0], out float value))
                 {
                     ScriptRange = value;
-                    Debug.WriteLine($"{ScriptName}: Received new {nameof(ScriptRange)} value {value}");
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Received new {nameof(ScriptRange)} value {value}");
                 }
-                else Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as float");
+                else CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as float");
 
             }), false);
 
@@ -638,167 +87,149 @@ namespace HandlingEditor.Client
             {
                 if (args.Count < 1)
                 {
-                    Debug.WriteLine($"{ScriptName}: Missing bool argument");
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Missing bool argument");
                     return;
                 }
 
                 if (bool.TryParse(args[0], out bool value))
                 {
-                    debug = value;
-                    Debug.WriteLine($"{ScriptName}: Received new {nameof(debug)} value {value}");
+                    Debug = value;
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Received new {nameof(Debug)} value {value}");
                 }
-                else Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as bool");
+                else CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as bool");
 
             }), false);
 
             RegisterCommand("handling_decorators", new Action<int, dynamic>((source, args) =>
             {
                 if (args.Count < 1)
-                    PrintDecorators(currentVehicle);
+                    PrintDecorators(CurrentVehicle);
                 else
                 {
                     if (int.TryParse(args[0], out int value))
                         PrintDecorators(value);
-                    else Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as int");
+                    else CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error parsing {args[0]} as int");
                 }
 
             }), false);
 
             RegisterCommand("handling_print", new Action<int, dynamic>((source, args) =>
             {
-                PrintVehiclesWithDecorators(vehicles);
+                PrintVehiclesWithDecorators(Vehicles);
             }), false);
 
             RegisterCommand("handling_preset", new Action<int, dynamic>((source, args) =>
             {
-                if (currentPreset != null)
-                    Debug.WriteLine(currentPreset.ToString());
+                if (CurrentPreset != null)
+                    CitizenFX.Core.Debug.WriteLine(CurrentPreset.ToString());
                 else
-                    Debug.WriteLine($"{ScriptName}: Current preset doesn't exist");
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Current preset doesn't exist");
             }), false);
 
             RegisterCommand("handling_xml", new Action<int, dynamic>((source, args) =>
             {
-                if (currentPreset != null)
-                    Debug.WriteLine(GetXmlFromPreset(currentPreset).OuterXml);
+                if (CurrentPreset != null)
+                    CitizenFX.Core.Debug.WriteLine(GetXmlFromPreset(CurrentPreset).OuterXml);
                 else
-                    Debug.WriteLine($"{ScriptName}: Current preset doesn't exist");
+                    CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Current preset doesn't exist");
             }), false);
 
+            HandlingMenu.PersonalPresetApplied += async (sender, name) =>
+            {
+                string key = $"{kvpPrefix}{name}";
+                string value = GetResourceKvpString(key);
+                if (value != null)
+                {
+                    value = Helpers.RemoveByteOrderMarks(value);
+                    XmlDocument doc = new XmlDocument();
+                    doc.LoadXml(value);
+                    var handling = doc["Item"];
+                    GetPresetFromXml(handling, CurrentPreset);
+
+                    Screen.ShowNotification($"{ScriptName}: Personal preset ~b~{name}~w~ applied");
+                }
+                else
+                    Screen.ShowNotification($"{ScriptName}: ~r~ERROR~w~ Personal preset ~b~{name}~w~ corrupted");
+
+                await Delay(1000);
+            };
+
+            HandlingMenu.ServerPresetApplied += async (sender, name) =>
+            {
+                string key = name;
+                if (ServerPresets.ContainsKey(key))
+                {
+                    foreach (var field in ServerPresets[key].Fields.Keys)
+                    {
+                        if (CurrentPreset.Fields.ContainsKey(field))
+                        {
+                            CurrentPreset.Fields[field] = ServerPresets[key].Fields[field];
+                        }
+                        else CitizenFX.Core.Debug.Write($"Missing {field} field in currentPreset");
+                    }
+
+                    Screen.ShowNotification($"{ScriptName}: Server preset ~b~{key}~w~ applied");
+                }
+                else
+                    Screen.ShowNotification($"{ScriptName}: ~r~ERROR~w~ Server preset ~b~{key}~w~ corrupted");
+
+                await Delay(1000);
+            };
+
+            HandlingMenu.ResetButtonPressed += async (sender, args) =>
+            {
+                CurrentPreset.Reset();
+                RefreshVehicleUsingPreset(CurrentVehicle, CurrentPreset);
+                RemoveDecorators(CurrentVehicle);
+
+                await Delay(200);
+                InvalidatedGUI(this, EventArgs.Empty);
+            };
+
             Tick += UpdateCurrentVehicle;
-            Tick += MenuTask;
             Tick += ScriptTask;
         }
+
+        #endregion
 
         #region TASKS
 
         /// <summary>
-        /// Updates the <see cref="currentVehicle"/> and the <see cref="currentPreset"/>
+        /// Updates the <see cref="CurrentVehicle"/> and the <see cref="CurrentPreset"/>
         /// </summary>
         /// <returns></returns>
         private async Task UpdateCurrentVehicle()
         {
-            playerPed = PlayerPedId();
+            PlayerPed = PlayerPedId();
 
-            if (IsPedInAnyVehicle(playerPed, false))
+            if (IsPedInAnyVehicle(PlayerPed, false))
             {
-                int vehicle = GetVehiclePedIsIn(playerPed, false);
+                int vehicle = GetVehiclePedIsIn(PlayerPed, false);
 
-                if (VehiclesPermissions.IsVehicleAllowed(vehicle) && GetPedInVehicleSeat(vehicle, -1) == playerPed && !IsEntityDead(vehicle))
+                if (VehiclesPermissions.IsVehicleAllowed(vehicle) && GetPedInVehicleSeat(vehicle, -1) == PlayerPed && !IsEntityDead(vehicle))
                 {
                     // Update current vehicle and get its preset
-                    if (vehicle != currentVehicle)
+                    if (vehicle != CurrentVehicle)
                     {
-                        currentVehicle = vehicle;
-                        currentPreset = CreateHandlingPreset(currentVehicle);
-                        InitialiseMenu();
+                        CurrentVehicle = vehicle;
+                        CurrentPreset = CreateHandlingPreset(CurrentVehicle);
+                        InvalidatedGUI.Invoke(this, EventArgs.Empty);
                     }
                 }
                 else
                 {
                     // If current vehicle isn't a car or player isn't driving current vehicle or vehicle is dead
-                    currentVehicle = -1;
-                    currentPreset = null;
+                    CurrentVehicle = -1;
+                    CurrentPreset = null;
                 }
             }
             else
             {
                 // If player isn't in any vehicle
-                currentVehicle = -1;
-                currentPreset = null;
+                CurrentVehicle = -1;
+                CurrentPreset = null;
             }
-        }
-
-        /// <summary>
-        /// The GUI task of the script
-        /// </summary>
-        /// <returns></returns>
-        private async Task MenuTask()
-        {
-            if(_menuPool != null)
-            {
-                _menuPool.ProcessMenus();
-
-                if (_menuPool.IsAnyMenuOpen())
-                    DisableControls();
-
-                if (currentVehicle != -1 && currentPreset != null)
-                {
-                    if (IsControlJustPressed(1, toggleMenu)/* || IsDisabledControlJustPressed(1, toggleMenu)*/) // TOGGLE MENU VISIBLE
-                    {
-                        if (_menuPool.IsAnyMenuOpen())
-                            _menuPool.CloseAllMenus();
-                        else
-                        {
-                            if(!EditorMenu.Visible)
-                                EditorMenu.Visible = true;
-                        }
-                            
-                    }
-
-                    if (PersonalPresetsMenu.Visible)
-                    {
-                        DisableControlAction(1, 75, true); // INPUT_VEH_EXIT - Y
-                        DisableControlAction(1, 37, true); // INPUT_SELECT_WEAPON - X
-
-                        if (IsControlJustPressed(1, 179))
-                        {
-                            string name = await GetOnScreenString("");
-                            if (!string.IsNullOrEmpty(name))
-                            {
-                                SavePresetAsKVP(name, currentPreset);
-                                UpdatePersonalPresetsMenu();
-                                Screen.ShowNotification($"{ScriptName}: Personal preset ~g~{name}~w~ saved");
-                            }
-                            else
-                                Screen.ShowNotification($"{ScriptName}: Invalid string.");
-                        }
-                        else if (IsControlJustPressed(1, 178))
-                        {
-                            if (PersonalPresetsMenu.MenuItems.Count > 0)
-                            {
-                                string kvpName = PersonalPresetsMenu.MenuItems[PersonalPresetsMenu.CurrentSelection].Text;
-                                string key = $"{kvpPrefix}{kvpName}";
-                                if (GetResourceKvpString(key) != null)
-                                {
-                                    DeleteResourceKvp(key);
-                                    UpdatePersonalPresetsMenu();
-
-                                    Screen.ShowNotification($"{ScriptName}: Personal preset ~r~{kvpName}~w~ deleted");
-                                }
-                            }
-                            else
-                                Screen.ShowNotification($"{ScriptName}: Nothing to delete.");
-                        }
-                    }
-
-                }
-                else
-                {
-                    _menuPool.CloseAllMenus();
-                }
-            }
-        }
+        }      
 
         /// <summary>
         /// The main task of the script
@@ -806,28 +237,28 @@ namespace HandlingEditor.Client
         /// <returns></returns>
         private async Task ScriptTask()
         {
-            currentTime = (GetGameTimer() - lastTime);
+            CurrentTime = (GetGameTimer() - LastTime);
 
             // Check if decorators needs to be updated
-            if (currentTime > timer)
+            if (CurrentTime > Timer)
             {
                 // Current vehicle could be updated each tick to show the edited fields live
                 // Check if current vehicle needs to be refreshed
-                if (currentVehicle != -1 && currentPreset != null)
+                if (CurrentVehicle != -1 && CurrentPreset != null)
                 {
-                    if (currentPreset.IsEdited)
-                        RefreshVehicleUsingPreset(currentVehicle, currentPreset);
+                    if (CurrentPreset.IsEdited)
+                        RefreshVehicleUsingPreset(CurrentVehicle, CurrentPreset);
                 }
 
-                if (currentVehicle != -1 && currentPreset != null)
-                    UpdateVehicleDecorators(currentVehicle, currentPreset);
+                if (CurrentVehicle != -1 && CurrentPreset != null)
+                    UpdateVehicleDecorators(CurrentVehicle, CurrentPreset);
 
-                vehicles = new VehicleEnumerable();
+                Vehicles = new VehicleEnumerable();
 
                 // Refreshes the iterated vehicles
-                RefreshVehicles(vehicles.Except(new List<int> { currentVehicle }));
+                RefreshVehicles(Vehicles.Except(new List<int> { CurrentVehicle }));
 
-                lastTime = GetGameTimer();
+                LastTime = GetGameTimer();
             }
         }
 
@@ -838,7 +269,7 @@ namespace HandlingEditor.Client
         /// <summary>
         /// Disable controls for controller to use the script with the controller
         /// </summary>
-        private void DisableControls()
+        public static void DisableControls()
         {
             DisableControlAction(1, 85, true); // INPUT_VEH_RADIO_WHEEL = DPAD - LEFT
             DisableControlAction(1, 74, true); // INPUT_VEH_HEADLIGHT = DPAD - RIGHT
@@ -846,6 +277,15 @@ namespace HandlingEditor.Client
             DisableControlAction(1, 27, true); // INPUT_PHONE = DPAD - UP
             DisableControlAction(1, 80, true); // INPUT_VEH_CIN_CAM = B
             DisableControlAction(1, 73, true); // INPUT_VEH_DUCK = A
+        }
+
+        /// <summary>
+        /// Disable controls for controller to use the script with the controller
+        /// </summary>
+        public static void DisableControls2()
+        {
+            DisableControlAction(1, 75, true); // INPUT_VEH_EXIT - Y
+            DisableControlAction(1, 37, true); // INPUT_SELECT_WEAPON - X
         }
 
         /// <summary>
@@ -863,11 +303,11 @@ namespace HandlingEditor.Client
                 string fieldName = item.Key;
                 dynamic fieldValue = item.Value;
 
-                var fieldsInfo = handlingInfo.FieldsInfo;
+                var fieldsInfo = HandlingInfo.FieldsInfo;
                 if (!fieldsInfo.ContainsKey(fieldName))
                 {
-                    if (debug)
-                        Debug.WriteLine($"{ScriptName}: No fieldInfo definition found for {fieldName}");
+                    if (Debug)
+                        CitizenFX.Core.Debug.WriteLine($"{ScriptName}: No fieldInfo definition found for {fieldName}");
                     continue;
                 }
 
@@ -882,8 +322,8 @@ namespace HandlingEditor.Client
                     {
                         SetVehicleHandlingFloat(vehicle, className, fieldName, fieldValue);
 
-                        if (debug)
-                            Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
+                        if (Debug)
+                            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
                     }
                 }
 
@@ -894,8 +334,8 @@ namespace HandlingEditor.Client
                     {
                         SetVehicleHandlingInt(vehicle, className, fieldName, fieldValue);
 
-                        if (debug)
-                            Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
+                        if (Debug)
+                            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
                     }
                 }
 
@@ -906,8 +346,8 @@ namespace HandlingEditor.Client
                     {
                         SetVehicleHandlingVector(vehicle, className, fieldName, fieldValue);
 
-                        if (debug)
-                            Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
+                        if (Debug)
+                            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {fieldValue}");
                     }
                 }
             }
@@ -919,7 +359,7 @@ namespace HandlingEditor.Client
         /// <param name="vehiclesList"></param>
         private void RefreshVehicles(IEnumerable<int> vehiclesList)
         {
-            Vector3 currentCoords = GetEntityCoords(playerPed, true);
+            Vector3 currentCoords = GetEntityCoords(PlayerPed, true);
 
             foreach (int entity in vehiclesList)
             {
@@ -939,7 +379,7 @@ namespace HandlingEditor.Client
         /// <param name="vehicle"></param>
         private void RefreshVehicleUsingDecorators(int vehicle)
         {
-            foreach (var item in handlingInfo.FieldsInfo.Where(a => a.Value.Editable))
+            foreach (var item in HandlingInfo.FieldsInfo.Where(a => a.Value.Editable))
             {
                 string fieldName = item.Key;
                 Type fieldType = item.Value.Type;
@@ -955,8 +395,8 @@ namespace HandlingEditor.Client
                         {
                             SetVehicleHandlingFloat(vehicle, className, fieldName, decorValue);
 
-                            if (debug)
-                                Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
+                            if (Debug)
+                                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
                         }
                     }
                 }
@@ -970,8 +410,8 @@ namespace HandlingEditor.Client
                         {
                             SetVehicleHandlingInt(vehicle, className, fieldName, decorValue);
 
-                            if (debug)
-                                Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
+                            if (Debug)
+                                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
                         }
                     }
                 }
@@ -997,8 +437,8 @@ namespace HandlingEditor.Client
                     {
                         SetVehicleHandlingVector(vehicle, className, fieldName, decorValue);
 
-                        if (debug)
-                            Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
+                        if (Debug)
+                            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: {fieldName} updated from {value} to {decorValue} for vehicle {vehicle}");
                     }
                 }
             }
@@ -1011,7 +451,7 @@ namespace HandlingEditor.Client
         /// <returns></returns>
         private bool HasDecorators(int vehicle)
         {
-            foreach (var item in handlingInfo.FieldsInfo)
+            foreach (var item in HandlingInfo.FieldsInfo)
             {
                 string fieldName = item.Key;
                 Type fieldType = item.Value.Type;
@@ -1032,7 +472,7 @@ namespace HandlingEditor.Client
         /// </summary>
         private void RegisterDecorators()
         {
-            foreach (var item in handlingInfo.FieldsInfo)
+            foreach (var item in HandlingInfo.FieldsInfo)
             {
                 string fieldName = item.Key;
                 Type type = item.Value.Type;
@@ -1070,7 +510,7 @@ namespace HandlingEditor.Client
         /// <param name="vehicle"></param>
         private void RemoveDecorators(int vehicle)
         {
-            foreach (var item in handlingInfo.FieldsInfo)
+            foreach (var item in HandlingInfo.FieldsInfo)
             {
                 string fieldName = item.Key;
                 Type fieldType = item.Value.Type;
@@ -1103,9 +543,9 @@ namespace HandlingEditor.Client
                 }
             }
 
-            if(debug)
+            if(Debug)
             {
-                Debug.WriteLine($"{ScriptName}: Removed all decorators on vehicle {vehicle}");
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Removed all decorators on vehicle {vehicle}");
             }
         }
 
@@ -1125,8 +565,8 @@ namespace HandlingEditor.Client
                 if (Math.Abs(currentValue - decorValue) > FloatPrecision)
                 {
                     DecorSetFloat(vehicle, name, currentValue);
-                    if (debug)
-                        Debug.WriteLine($"{ScriptName}: Updated decorator {name} updated from {decorValue} to {currentValue} for vehicle {vehicle}");
+                    if (Debug)
+                        CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Updated decorator {name} updated from {decorValue} to {currentValue} for vehicle {vehicle}");
                 }
             }
             else // Decorator doesn't exist, create it if required
@@ -1134,8 +574,8 @@ namespace HandlingEditor.Client
                 if (Math.Abs(currentValue - defaultValue) > FloatPrecision)
                 {
                     DecorSetFloat(vehicle, name, currentValue);
-                    if (debug)
-                        Debug.WriteLine($"{ScriptName}: Added decorator {name} with value {currentValue} to vehicle {vehicle}");
+                    if (Debug)
+                        CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Added decorator {name} with value {currentValue} to vehicle {vehicle}");
                 }
             }
         }
@@ -1156,8 +596,8 @@ namespace HandlingEditor.Client
                 if (currentValue != decorValue)
                 {
                     DecorSetInt(vehicle, name, currentValue);
-                    if (debug)
-                        Debug.WriteLine($"{ScriptName}: Updated decorator {name} updated from {decorValue} to {currentValue} for vehicle {vehicle}");
+                    if (Debug)
+                        CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Updated decorator {name} updated from {decorValue} to {currentValue} for vehicle {vehicle}");
                 }
             }
             else // Decorator doesn't exist, create it if required
@@ -1165,8 +605,8 @@ namespace HandlingEditor.Client
                 if (currentValue != defaultValue)
                 {
                     DecorSetInt(vehicle, name, currentValue);
-                    if (debug)
-                        Debug.WriteLine($"{ScriptName}: Added decorator {name} with value {currentValue} to vehicle {vehicle}");
+                    if (Debug)
+                        CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Added decorator {name} with value {currentValue} to vehicle {vehicle}");
                 }
             }
         }
@@ -1180,7 +620,7 @@ namespace HandlingEditor.Client
             foreach (var item in preset.Fields)
             {
                 string fieldName = item.Key;
-                Type fieldType = handlingInfo.FieldsInfo[fieldName].Type;
+                Type fieldType = HandlingInfo.FieldsInfo[fieldName].Type;
                 dynamic fieldValue = item.Value;
 
                 string defDecorName = $"{fieldName}_def";
@@ -1230,7 +670,7 @@ namespace HandlingEditor.Client
             Dictionary<string, dynamic> defaultFields = new Dictionary<string, dynamic>();
             Dictionary<string, dynamic> fields = new Dictionary<string, dynamic>();
             
-            foreach(var item in handlingInfo.FieldsInfo)
+            foreach(var item in HandlingInfo.FieldsInfo)
             {
                 string fieldName = item.Key;
                 string className = item.Value.ClassName;
@@ -1298,7 +738,7 @@ namespace HandlingEditor.Client
                 s.AppendLine($"{ScriptName}: Vehicle:{vehicle} netID:{netID}");
                 s.AppendLine("Decorators List:");
 
-                foreach (var item in handlingInfo.FieldsInfo)
+                foreach (var item in HandlingInfo.FieldsInfo)
                 {
                     string fieldName = item.Key;
                     Type fieldType = item.Value.Type;
@@ -1355,9 +795,9 @@ namespace HandlingEditor.Client
                         
                     }
                 }
-                Debug.Write(s.ToString());
+                CitizenFX.Core.Debug.Write(s.ToString());
             }
-            else Debug.WriteLine($"{ScriptName}: Can't find vehicle with handle {vehicle}");
+            else CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Can't find vehicle with handle {vehicle}");
         }
 
         /// <summary>
@@ -1367,7 +807,7 @@ namespace HandlingEditor.Client
         {
             IEnumerable<int> entities = vehiclesList.Where(entity => HasDecorators(entity));
 
-            Debug.WriteLine($"{ScriptName}: Vehicles with decorators: {entities.Count()}");
+            CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Vehicles with decorators: {entities.Count()}");
 
             StringBuilder s = new StringBuilder();
             foreach (var vehicle in entities)
@@ -1375,10 +815,10 @@ namespace HandlingEditor.Client
                 int netID = NetworkGetNetworkIdFromEntity(vehicle);      
                 s.AppendLine($"Vehicle:{vehicle} netID:{netID}");
             }
-            Debug.WriteLine(s.ToString());
+            CitizenFX.Core.Debug.WriteLine(s.ToString());
         }
 
-        private XmlDocument GetXmlFromPreset(HandlingPreset preset)
+        private static XmlDocument GetXmlFromPreset(HandlingPreset preset)
         {
             XmlDocument doc = new XmlDocument();
             XmlElement handlingItem = doc.CreateElement("Item");
@@ -1390,7 +830,7 @@ namespace HandlingEditor.Client
                 dynamic fieldValue = item.Value;
                 XmlElement field = doc.CreateElement(fieldName);
 
-                Type fieldType = handlingInfo.FieldsInfo[fieldName].Type;
+                Type fieldType = HandlingInfo.FieldsInfo[fieldName].Type;
                 if(fieldType == FieldType.FloatType)
                 {
                     var value = (float)fieldValue;
@@ -1423,7 +863,7 @@ namespace HandlingEditor.Client
             return doc;
         }
 
-        private void SavePresetAsKVP(string name, HandlingPreset preset)
+        public static void SavePresetAsKVP(string name, HandlingPreset preset)
         {
             string kvpName = $"{kvpPrefix}{name}";
             if(GetResourceKvpString(kvpName) != null)
@@ -1476,15 +916,15 @@ namespace HandlingEditor.Client
             try
             {
                 strings = LoadResourceFile(ResourceName, filename);
-                handlingInfo.ParseXml(strings);
-                var editableFields = handlingInfo.FieldsInfo.Where(a => a.Value.Editable);
-                Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {handlingInfo.FieldsInfo.Count} fields info, {editableFields.Count()} editable.");
+                HandlingInfo.ParseXml(strings);
+                var editableFields = HandlingInfo.FieldsInfo.Where(a => a.Value.Editable);
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {HandlingInfo.FieldsInfo.Count} fields info, {editableFields.Count()} editable.");
             }
             catch (Exception e)
             {
-                Debug.WriteLine(e.Message);
-                Debug.WriteLine(e.StackTrace);
-                Debug.WriteLine($"{ScriptName}: Error loading {filename}");
+                CitizenFX.Core.Debug.WriteLine(e.Message);
+                CitizenFX.Core.Debug.WriteLine(e.StackTrace);
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error loading {filename}");
             }
         }
 
@@ -1495,13 +935,13 @@ namespace HandlingEditor.Client
             {
                 strings = LoadResourceFile(ResourceName, filename);
                 VehiclesPermissions.ParseXml(strings);
-                Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {VehiclesPermissions.Classes.Count} class rules and {VehiclesPermissions.Vehicles.Count} vehicle rules");
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {VehiclesPermissions.Classes.Count} class rules and {VehiclesPermissions.Vehicles.Count} vehicle rules");
             }
             catch (Exception e)
             {
-                Debug.WriteLine(e.Message);
-                Debug.WriteLine(e.StackTrace);
-                Debug.WriteLine($"{ScriptName}: Error loading {filename}");
+                CitizenFX.Core.Debug.WriteLine(e.Message);
+                CitizenFX.Core.Debug.WriteLine(e.StackTrace);
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error loading {filename}");
             }
         }
 
@@ -1526,16 +966,16 @@ namespace HandlingEditor.Client
                         HandlingPreset preset = new HandlingPreset();
 
                         GetPresetFromXml(node, preset);
-                        serverPresets[name] = preset;
+                        ServerPresets[name] = preset;
                     }
                 }
-                Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {serverPresets.Count} server presets.");
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Loaded {filename}, found {ServerPresets.Count} server presets.");
             }
             catch (Exception e)
             {
-                Debug.WriteLine(e.Message);
-                Debug.WriteLine(e.StackTrace);
-                Debug.WriteLine($"{ScriptName}: Error loading {filename}");
+                CitizenFX.Core.Debug.WriteLine(e.Message);
+                CitizenFX.Core.Debug.WriteLine(e.StackTrace);
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Error loading {filename}");
             }
         }
 
@@ -1546,26 +986,26 @@ namespace HandlingEditor.Client
             {
                 strings = LoadResourceFile(ResourceName, filename);
 
-                Debug.WriteLine($"{ScriptName}: Loaded settings from {filename}");
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Loaded settings from {filename}");
             }
             catch (Exception e)
             {
-                Debug.WriteLine($"{ScriptName}: Impossible to load {filename}");
-                Debug.WriteLine(e.StackTrace);
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Impossible to load {filename}");
+                CitizenFX.Core.Debug.WriteLine(e.StackTrace);
             }
             finally
             {
                 Config config = new Config(strings);
 
-                toggleMenu = config.GetIntValue("toggleMenu", toggleMenu);
+                ToggleMenu = config.GetIntValue("toggleMenu", ToggleMenu);
                 FloatStep = config.GetFloatValue("FloatStep", FloatStep);
                 ScriptRange = config.GetFloatValue("ScriptRange", ScriptRange);
-                timer = config.GetLongValue("timer", timer);
-                debug = config.GetBoolValue("debug", debug);
-                screenPosX = config.GetFloatValue("screenPosX", screenPosX);
-                screenPosY = config.GetFloatValue("screenPosY", screenPosY);
+                Timer = config.GetLongValue("timer", Timer);
+                Debug = config.GetBoolValue("debug", Debug);
+                ScreenPosX = config.GetFloatValue("screenPosX", ScreenPosX);
+                ScreenPosY = config.GetFloatValue("screenPosY", ScreenPosY);
 
-                Debug.WriteLine($"{ScriptName}: Settings {nameof(timer)}={timer} {nameof(debug)}={debug} {nameof(ScriptRange)}={ScriptRange}");
+                CitizenFX.Core.Debug.WriteLine($"{ScriptName}: Settings {nameof(Timer)}={Timer} {nameof(Debug)}={Debug} {nameof(ScriptRange)}={ScriptRange}");
             }
         }
